# Align AccessListMember.Clone method