--- a/api/types/accesslist/member.go
+++ b/api/types/accesslist/member.go
@@ -125,9 +125,23 @@ func (a *AccessListMember) MatchSearch(values []string) bool {
 
 // Clone returns a copy of the member.
 func (a *AccessListMember) Clone() *AccessListMember {
-	var copy *AccessListMember
-	utils.StrictObjectToStruct(a, &copy)
-	return copy
+	if a == nil {
+		return nil
+	}
+	return &AccessListMember{
+		ResourceHeader: *a.ResourceHeader.Clone(),
+		Spec: AccessListMemberSpec{
+			AccessList:       a.Spec.AccessList,
+			Name:             a.Spec.Name,
+			Title:            a.Spec.Title,
+			Joined:           a.Spec.Joined,
+			Expires:          a.Spec.Expires,
+			Reason:           a.Spec.Reason,
+			AddedBy:          a.Spec.AddedBy,
+			IneligibleStatus: a.Spec.IneligibleStatus,
+			MembershipKind:   a.Spec.MembershipKind,
+		},
+	}
 }
 
 // IsExpired checks if the access list member is expired based on the current time.