### What

Access list members count can be > 10M. The json clone doesn't scale for this case. 

### Before: 

```
cpu: Apple M3 Max
BenchmarkAccessListClone
BenchmarkAccessListClone/count=100
BenchmarkAccessListClone/count=100-14         	   20560	     57787 ns/op	  243458 B/op	     600 allocs/op
BenchmarkAccessListClone/count=1000
BenchmarkAccessListClone/count=1000-14        	    2056	    584011 ns/op	 2434603 B/op	    6007 allocs/op
BenchmarkAccessListClone/count=10000
BenchmarkAccessListClone/count=10000-14       	     198	   6000901 ns/op	24346672 B/op	   60080 allocs/op
BenchmarkAccessListClone/count=100000
BenchmarkAccessListClone/count=100000-14      	      18	  60888808 ns/op	243399482 B/op	  600710 allocs/op
PASS

```


### After: 

```
BenchmarkAccessListClone
BenchmarkAccessListClone/count=100
BenchmarkAccessListClone/count=100-14         	 7814946	       136.4 ns/op	       0 B/op	       0 allocs/op
BenchmarkAccessListClone/count=1000
BenchmarkAccessListClone/count=1000-14        	  886110	      1393 ns/op	       0 B/op	       0 allocs/op
BenchmarkAccessListClone/count=10000
BenchmarkAccessListClone/count=10000-14       	   78871	     13630 ns/op	       0 B/op	       0 allocs/op
BenchmarkAccessListClone/count=100000
BenchmarkAccessListClone/count=100000-14      	    8863	    133654 ns/op	       0 B/op	       0 allocs/op
PASS

```