@kopiczko 2025-09-02
Looks good, but let's do an equivalent test as in https://github.com/gravitational/teleport/pull/58592